Feature: test udpdates and error handling
